アカウント設定のコンポーネントの分離
